Split finder into separate classes.
