+++ packages/config/src/types.ts
@@ -8,14 +8,16 @@ export type Loader<T> = (path: Path, pkg: PackageStructure) => Promise<Partial<T
 
 export type Handler<T> = (prev: T, next: T) => T | undefined | Promise<T | undefined>;
 
-export interface ConfigFile<T> {
-  config: Partial<T>;
+export interface File {
   path: Path;
 }
 
-export interface IgnoreFile {
+export interface ConfigFile<T> extends File {
+  config: Partial<T>;
+}
+
+export interface IgnoreFile extends File {
   ignore: string[];
-  path: Path;
 }
 
 export interface ProcessedConfig<T> {
@@ -23,13 +25,17 @@ export interface ProcessedConfig<T> {
   files: ConfigFile<T>[];
 }
 
-export interface FinderOptions<T> {
+export interface ConfigFinderOptions<T> {
   extendsSetting?: string;
   extensions?: ExtType[];
   includeEnv?: boolean;
   loaders?: { [K in LoaderType]: Loader<T> };
   name: string;
   overridesSetting?: string;
+}
+
+export interface IgnoreFinderOptions {
+  name: string;
 }
 
 export type FileGlob = string | string[];
